Reorganise search, and everything
So, I decided we needed to refactor stuff...again.

The search folder (inside assets/) is the best example we have so far.
But I think it still needs a wee bit of reorganisation. But it's much
better than it was.

Everything else needs a lot more work done, but stuff is reorganised
slightly (views folder will eventually disappear, but is at least with
all the other assets now).